server/requirements.txt
appdirs==1.4.3
Django==2.1
psycopg2
djangorestframework==3.9
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
django-cors-headers